+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/request/RequestArgs.kt
@@ -1,0 +1,8 @@
+package co.electriccoin.zcash.ui.screen.request
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class RequestArgs(
+    val addressType: Int
+)