+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/request/view/RequestAmountView.kt
@@ -122,11 +122,7 @@ private fun RequestAmountWithMainFiatView(
         val fiatText =
             buildAnnotatedString {
                 withStyle(style = SpanStyle(color = ZashiColors.Text.textQuaternary)) {
-                    append(
-                        state.exchangeRateState.currencyConversion
-                            ?.fiatCurrency
-                            ?.symbol
-                    )
+                    append(state.exchangeRateState.expectedCurrency.symbol)
                 }
                 append("\u2009") // Add an extra thin space between the texts
                 withStyle(style = SpanStyle(color = ZashiColors.Text.textPrimary)) {
@@ -230,11 +226,7 @@ private fun RequestAmountWithMainZecView(
             val fiatText =
                 buildAnnotatedString {
                     withStyle(style = SpanStyle(color = ZashiColors.Text.textQuaternary)) {
-                        append(
-                            state.exchangeRateState.currencyConversion
-                                ?.fiatCurrency
-                                ?.symbol
-                        )
+                        append(state.exchangeRateState.expectedCurrency.symbol)
                     }
                     append(" ") // Add an extra space between the texts
                     withStyle(style = SpanStyle(color = ZashiColors.Text.textPrimary)) {